#1713 change area_code, location_id to LONG type

+++ src/main/java/at/rtr/rmbt/repository/impl/RadioSignalRepositoryImpl.java
@@ -161,7 +161,7 @@ public class RadioSignalRepositoryImpl implements RadioSignalRepository {
                                 rssnr,
                                 rsSignal.getString("cat_technology"),
                                 (rsSignal.getObject("location_id", Long.class) == null ? null : rsSignal.getObject("location_id", Long.class)),
-                                rsSignal.getObject("area_code", Integer.class),
+                                rsSignal.getObject("area_code", Long.class),
                                 rsSignal.getObject("primary_scrambling_code", Integer.class),
                                 rsSignal.getObject("channel_number", Integer.class),
                                 rsSignal.getObject("timing_advance", Integer.class));
